+++ 07-problems-level-3/p07.cpp
@@ -26,7 +26,7 @@ void printMatrix(int matrix[3][3], short matrixRows, short matrixColumns, string
 	cout << prompt;
 	for (short i = 0; i < matrixRows; i++)
 	{
-		for (int j = 0; j < matrixColumns; j++)
+		for (short j = 0; j < matrixColumns; j++)
 		{
 			cout << left << setw(3) << matrix[i][j] << "  ";
 		}
@@ -38,7 +38,7 @@ void fillMatrixWithRandomNumbers(int matrix[3][3],short matrixRows, short matrix
 {
 	for (short i = 0; i < matrixRows; i++)
 	{
-		for (int j = 0; j < matrixColumns; j++)
+		for (short j = 0; j < matrixColumns; j++)
 		{
 			matrix[i][j] = getRandomNumber(1, 20);
 		}
@@ -57,7 +57,7 @@ void fillMatrixWithOrderedNumbers(int matrix[3][3], short matrixRows, short matr
 	short number = 0;
 	for (int i = 0; i < matrixRows; i++)
 	{
-		for (int j = 0; j < matrixColumns; j++)
+		for (short j = 0; j < matrixColumns; j++)
 		{
 			matrix[i][j] = ++number;
 		}
@@ -68,7 +68,7 @@ void turnMatrixToTranspose(int matrix[3][3], short matrixRows, short matrixColum
 {
 	for (int i = 0; i < matrixRows; i++)
 	{
-		for (int j = i+1; j < matrixColumns; j++)
+		for (short j = i+1; j < matrixColumns; j++)
 		{
 			if (i == 0 && j == 0)
 				continue;
@@ -81,7 +81,7 @@ void turnMatrixToTranspose(int matrix1[3][3], int matrix2[3][3], short matrixRow
 {
 	for (int i = 0; i < matrixRows; i++)
 	{
-		for (int j = 0; j < matrixColumns; j++)
+		for (short j = 0; j < matrixColumns; j++)
 		{
 			matrix2[i][j] = matrix1[j][i];
 		}
